Serialize & deserialize "inline" flag in LTO cgraphs
When LTO runs, cgraph_edge->call_stmt is NULL. It is not reconstructed
when reading the cgraph data from .o files.
So I store that bit on the cgraph_edge itself, and add it to the
serialized format.
When writing the "inline" bit - if the call_stmt is available, I use
the "inline" flag from the gimple call.